Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,7 +5,7 @@ int main(){
     int c;
     float Average;
     scanf("%d%d%d",&a,&b,&c);
-    Average= (a+b+c)/3;
+    Average= (float)(a+b+c)/3;
     printf("Average: %.2f",Average);
     return 0;
 }
